+++ fixtures/tests/ca18b090-03ad-4bef-acd3-52600f8e62eb.run.full.json
@@ -5,7 +5,7 @@
         "fields": {
             "created_date": "2019-12-11T22:53:02.012Z",
             "modified_date": "2020-01-14T23:03:35.530Z",
-            "name": "ROSLIN 09670_D, 22 of 54 (12/11/2019, 22:53:02)",
+            "name": "ARGOS 09670_D, 22 of 54 (12/11/2019, 22:53:02)",
             "app": "cb5d793b-e650-4b7d-bfcd-882858e29cc5",
             "status": 4,
             "execution_id": "df43cc77-bed4-4233-9817-6d12e25d9e1c",
