+++ runner/fixtures/runner.pipeline.json
@@ -33,8 +33,8 @@
         "fields": {
             "created_date": "2019-11-18T17:46:45.118Z",
             "modified_date": "2019-12-05T01:12:39.854Z",
-            "name": "roslin",
-            "github": "https://github.com/mskcc/roslin-cwl",
+            "name": "argos",
+            "github": "https://github.com/mskcc/argos-cwl",
             "version": "1.0.0-rc5",
             "entrypoint": "workflows/pair-workflow-sv.cwl",
             "output_file_group": "a975f490-1b02-4575-abae-a4f8e3667733",
@@ -63,7 +63,7 @@
             "created_date": "2019-11-18T17:46:45.118Z",
             "modified_date": "2019-12-05T01:12:39.854Z",
             "name": "copy_outputs",
-            "github": "https://github.com/mskcc/roslin-cwl",
+            "github": "https://github.com/mskcc/argos-cwl",
             "version": "1.0.0",
             "entrypoint": "workflows/copy_outputs.cwl",
             "output_file_group": "a975f490-1b02-4575-abae-a4f8e3667733",
