+++ runner/tests/operator/copy_outputs_operator/test_copy_outputs_operator.py
@@ -1,5 +1,5 @@
 """
-Test for constructing Roslin pair and jobs
+Test for constructing Argos pair and jobs
 """
 import os
 import json
